requirements.txt
_libgcc_mutex=0.1=main
absl-py=0.9.0
astor=0.8.1
ca-certificates==2020.1.1
certifi==2019.11.28
cycler==0.10.0
django==2.2.5
django-crispy-forms==1.7.2
freetype
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
jpeg==9b==h024ee3a_2
keras-applications==1.0.8==pypi_0
keras-preprocessing==1.1.0==pypi_0
kiwisolver==1.1.0==pypi_0
ld_impl_linux-64==2.33.1==h53a641e_7
libedit==3.1.20181209==hc058e9b_0
libffi==3.2.1==hd88cf55_4
libgcc-ng==9.1.0==hdf63c60_0
libpng==1.6.37==hbc83047_0
libstdcxx-ng==9.1.0==hdf63c60_0
libtiff==4.1.0==h2733197_0
markdown==3.2.1==pypi_0
matplotlib==3.2.0==pypi_0
ncurses==6.2==he6710b0_0
numpy==1.18.1==pypi_0
olefile==0.46==py37_0
opencv-contrib-python==4.2.0.32==pypi_0
openssl==1.1.1d==h7b6447c_4
opt-einsum==3.2.0==pypi_0
pillow==5.3.0==py37h34e0f95_0
pip==20.0.2==py37_1
protobuf==3.11.3==pypi_0
pyparsing==2.4.6==pypi_0
python==3.7.6==h0371630_2
python-dateutil==2.8.1==pypi_0
pytz==2018.4==pypi_0
readline==7.0==h7b6447c_5
reportlab==3.5.26==py37he686d34_0
setuptools==46.0.0==py37_0
six==1.14.0==pypi_0
sqlite==3.31.1==h7b6447c_0
sqlparse==0.3.1==pypi_0
tensorboard==1.15.0==pypi_0
tensorflow==1.15.0==pypi_0
tensorflow-estimator==1.15.1==pypi_0
termcolor==1.1.0==pypi_0
tk==8.6.8==hbc83047_0
werkzeug==1.0.0==pypi_0
wheel==0.34.2==py37_0
wrapt==1.12.1==pypi_0
xz==5.2.4==h14c3975_4
zlib==1.2.11==h7b6447c_3
zstd==1.3.7==h0b5b093_0